Chg: If file_get_contents is false on yydecode, return PEAR Error.

+++ nzedb/controllers/NNTP.php
@@ -1001,6 +1001,9 @@ class NNTP extends Net_NNTP_Client
 					$this->yEncSilence
 				);
 				$data = file_get_contents($ouFile);
+				if ($data === false) {
+					return $this->throwError('Error getting data from yydecode.');
+				}
 				unlink($inFile);
 				unlink($ouFile);
 			}
